Cleaned up reply chain logic
